+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.0.26
+
+- Change API update from 30 seconds to 5 minutes, will update also when action is being performed
+
 ## v3.0.25
 
 ### Breaking Change!